2_autoinstrument_jaeger/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Deprecated==1.2.13
Flask==2.1.2
Flask-Cors
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
opentelemetry-api==1.11.1
opentelemetry-exporter-jaeger
opentelemetry-exporter-jaeger-proto-grpc
opentelemetry-exporter-jaeger-thrift
opentelemetry-distro==0.30b1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
